:recycle: #21 Refactor: 디렉토리 정리
