+++ src/main/java/com/umc/TheGoods/config/springSecurity/utils/JwtTokenFilter.java
@@ -1,7 +1,7 @@
 package com.umc.TheGoods.config.springSecurity.utils;
 
-import com.umc.TheGoods.service.Member.MemberCommandServiceImpl;
-import com.umc.TheGoods.web.dto.Member.MemberDetail;
+import com.umc.TheGoods.service.MemberService.MemberCommandServiceImpl;
+import com.umc.TheGoods.web.dto.member.MemberDetail;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
