+++ src/main/java/com/umc/TheGoods/repository/member/TermRepository.java
@@ -1,4 +1,4 @@
-package com.umc.TheGoods.repository;
+package com.umc.TheGoods.repository.member;
 
 import com.umc.TheGoods.domain.member.Term;
 import org.springframework.data.jpa.repository.JpaRepository;
