+++ src/main/java/com/umc/TheGoods/service/MemberService/MemberCommandServiceImpl.java
@@ -1,21 +1,21 @@
-package com.umc.TheGoods.service.Member;
+package com.umc.TheGoods.service.MemberService;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.umc.TheGoods.apiPayload.code.status.ErrorStatus;
 import com.umc.TheGoods.apiPayload.exception.handler.MemberHandler;
-import com.umc.TheGoods.converter.Member.MemberConverter;
+import com.umc.TheGoods.converter.member.MemberConverter;
 import com.umc.TheGoods.domain.item.Category;
 import com.umc.TheGoods.domain.mapping.member.MemberCategory;
 import com.umc.TheGoods.domain.mapping.member.MemberTerm;
 import com.umc.TheGoods.domain.member.Member;
 import com.umc.TheGoods.domain.member.PhoneAuth;
 import com.umc.TheGoods.domain.member.Term;
-import com.umc.TheGoods.repository.CategoryRepository;
-import com.umc.TheGoods.repository.MemberRepository;
-import com.umc.TheGoods.repository.PhoneAuthRepository;
-import com.umc.TheGoods.repository.TermRepository;
-import com.umc.TheGoods.web.dto.Member.MemberRequestDTO;
+import com.umc.TheGoods.repository.member.CategoryRepository;
+import com.umc.TheGoods.repository.member.MemberRepository;
+import com.umc.TheGoods.repository.member.PhoneAuthRepository;
+import com.umc.TheGoods.repository.member.TermRepository;
+import com.umc.TheGoods.web.dto.member.MemberRequestDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
