+++ src/main/java/com/umc/TheGoods/web/controller/OrderController.java
@@ -3,12 +3,12 @@ package com.umc.TheGoods.web.controller;
 import com.umc.TheGoods.apiPayload.ApiResponse;
 import com.umc.TheGoods.apiPayload.code.status.ErrorStatus;
 import com.umc.TheGoods.apiPayload.exception.handler.MemberHandler;
-import com.umc.TheGoods.converter.orders.OrderConverter;
+import com.umc.TheGoods.converter.order.OrderConverter;
 import com.umc.TheGoods.domain.member.Member;
 import com.umc.TheGoods.domain.order.Orders;
-import com.umc.TheGoods.service.Member.MemberQueryService;
+import com.umc.TheGoods.service.MemberService.MemberQueryService;
 import com.umc.TheGoods.service.OrderService.OrderCommandService;
-import com.umc.TheGoods.web.dto.Member.MemberDetail;
+import com.umc.TheGoods.web.dto.member.MemberDetail;
 import com.umc.TheGoods.web.dto.order.OrderRequest;
 import com.umc.TheGoods.web.dto.order.OrderResponse;
 import io.swagger.v3.oas.annotations.Operation;
